requirements.txt
asgiref==3.6.0
crispy-bootstrap5==0.7
Django==4.1.7
django-crispy-forms==2.0
django-filter==23.1
django-phonenumber-field==7.0.2
django-widget-tweaks==1.4.12
phonenumberslite==8.13.8
Pillow==9.4.0
sqlparse==0.4.3
